--- bin/test_01.py
@@ -1,1 +1,0 @@
-print("Hello world in Python file")
